+++ packages/app-main/src/main/app/getAppDetails.ts
@@ -1,4 +1,4 @@
-import type { AppDetails } from '@ter/app-common/apis/app';
+import type { AppDetails } from '@app/common/apis/app';
 import { app } from 'electron';
 
 export function getAppDetails(): AppDetails {