+++ packages/app-main/src/main/window/types.ts
@@ -1,4 +1,4 @@
-import type { WindowType } from '@ter/app-common/apis/app';
+import type { WindowType } from '@app/common/apis/app';
 
 export interface WindowOptions {
     onClosed?: (windowId: number) => void | Promise<void>;