+++ packages/app-renderer/src/apis/exposes.ts
@@ -1,5 +1,5 @@
-import type { AppRendererAPI } from '@ter/app-common/apis/app';
-import type { LoggerRendererAPI } from '@ter/app-common/apis/logger';
+import type { AppRendererAPI } from '@app/common/apis/app';
+import type { LoggerRendererAPI } from '@app/common/apis/logger';
 
 declare const window: {
     /* eslint-disable @typescript-eslint/naming-convention */