+++ packages/app-main/src/main/configuration/store.ts
@@ -4,10 +4,12 @@ import type { IWindowState } from './window';
 
 export enum ConfigurationKey {
     MAIN_WINDOW_STATE = 'mainWindowState',
+    RESTORE_LAST_WINDOW_STATE = 'restoreLastWindowState',
 }
 
 interface IAppConfiguration {
     [ConfigurationKey.MAIN_WINDOW_STATE]?: IWindowState;
+    [ConfigurationKey.RESTORE_LAST_WINDOW_STATE]?: boolean;
 }
 
 export const store = new Store<IAppConfiguration>({