+++ packages/app-main/src/main/configuration/window.ts
@@ -21,64 +21,37 @@ const keyMap: Record<WindowType, WindowStateKeys> = {
     [WindowType.MAIN]: ConfigurationKey.MAIN_WINDOW_STATE,
 };
 
-function checkWindowVisible(state: IWindowState) {
-    const checkWindowPosition = (bounds: Electron.Rectangle) => {
-        return (
-            state.x >= bounds.x &&
-            state.y >= bounds.y &&
-            state.x + state.width <= bounds.x + bounds.width &&
-            state.y + state.height <= bounds.y + bounds.height
-        );
-    };
-
-    const allDisplays = screen.getAllDisplays();
-    const visible = allDisplays.some((display) => checkWindowPosition(display.bounds));
-    return visible;
-}
-
-function getWindowState(type: WindowType): IWindowState | null {
-    const prevState = store.get(keyMap[type]);
-    if (prevState) {
-        const visible = checkWindowVisible(prevState);
-        if (visible) {
-            return prevState;
-        }
-    }
-
-    return null;
-}
-
 interface IWindowStateKeeperOptions {
-    width?: number;
-    height?: number;
-    maximized?: boolean;
-    fullScreen?: boolean;
-    ignorePrevious?: boolean;
+    defaultWidth?: number;
+    defaultHeight?: number;
+    defaultMaximized?: boolean;
+    defaultFullScreen?: boolean;
 }
 
 export class WindowStateKeeper {
     private readonly _type: WindowType;
-
+    enabled: boolean;
     private readonly _state: IWindowState;
     private _stateChangeTimer?: NodeJS.Timeout;
     private _windowRef?: BrowserWindow;
 
-    private readonly _defaultWidth: number;
-    private readonly _defaultHeight: number;
-    private readonly _defaultMaximized: boolean;
-    private readonly _defaultFullScreen: boolean;
-
-    constructor(type: WindowType, options?: IWindowStateKeeperOptions) {
+    constructor(type: WindowType, enabled: boolean, options?: IWindowStateKeeperOptions) {
         this._type = type;
+        this.enabled = enabled;
 
-        this._defaultWidth = options?.width ?? 1280;
-        this._defaultHeight = options?.height ?? 720;
-        this._defaultMaximized = options?.maximized ?? false;
-        this._defaultFullScreen = options?.fullScreen ?? false;
+        const { workAreaSize } = screen.getPrimaryDisplay();
+        const defaultWidth = options?.defaultWidth ?? 1280;
+        const defaultHeight = options?.defaultHeight ?? 720;
+        this._state = {
+            x: Math.round((workAreaSize.width - defaultWidth) / 2),
+            y: Math.round((workAreaSize.height - defaultHeight) / 2),
+            width: defaultWidth,
+            height: defaultHeight,
+            maximized: options?.defaultMaximized ?? false,
+            fullScreen: options?.defaultFullScreen ?? false,
+        };
 
-        this._state = this._getDefaultWindowState();
-
-        if (!options?.ignorePrevious) {
+        if (this.enabled) {
             const prevState = getWindowState(type);
             if (prevState) {
                 this._state = prevState;
@@ -110,19 +83,6 @@ export class WindowStateKeeper {
         return this._state.fullScreen;
     }
 
-    private _getDefaultWindowState(): IWindowState {
-        const { workAreaSize } = screen.getPrimaryDisplay();
-
-        return {
-            x: Math.round((workAreaSize.width - this._defaultWidth) / 2),
-            y: Math.round((workAreaSize.height - this._defaultHeight) / 2),
-            width: this._defaultWidth,
-            height: this._defaultHeight,
-            maximized: this._defaultMaximized,
-            fullScreen: this._defaultFullScreen,
-        };
-    }
-
     private _updateState = () => {
         const window = this._windowRef;
         if (!window) return;
@@ -143,7 +103,11 @@ export class WindowStateKeeper {
     };
 
     private _saveState() {
-        store.set(keyMap[this._type], this._state);
+        if (this.enabled) {
+            store.set(keyMap[this._type], this._state);
+        } else {
+            store.delete(keyMap[this._type]);
+        }
     }
 
     private _handleStateChange = () => {
@@ -166,7 +130,7 @@ export class WindowStateKeeper {
         window.on('move', this._handleStateChange);
         window.on('resize', this._handleStateChange);
         window.on('close', this._handleClose);
-        window.on('closed', this._handleClosed);
+        window.once('closed', this._handleClosed);
         this._windowRef = window;
     }
 
@@ -178,7 +142,33 @@ export class WindowStateKeeper {
             this._stateChangeTimer = undefined;
         }
         this._windowRef?.removeListener('close', this._handleClose);
-        this._windowRef?.removeListener('closed', this._handleClosed);
         this._windowRef = undefined;
     }
+}
+
+function getWindowState(type: WindowType): IWindowState | null {
+    const prevState = store.get(keyMap[type]);
+    if (prevState) {
+        const visible = checkWindowVisible(prevState);
+        if (visible) {
+            return prevState;
+        }
+    }
+
+    return null;
+}
+
+function checkWindowVisible(state: IWindowState) {
+    const checkWindowPosition = (bounds: Electron.Rectangle) => {
+        return (
+            state.x >= bounds.x &&
+            state.y >= bounds.y &&
+            state.x + state.width <= bounds.x + bounds.width &&
+            state.y + state.height <= bounds.y + bounds.height
+        );
+    };
+
+    const allDisplays = screen.getAllDisplays();
+    const visible = allDisplays.some((display) => checkWindowPosition(display.bounds));
+    return visible;
 }